refactor: consolidate KDA shared FLA helpers
(cherry picked from commit b37dbf156b218350d847f8fda657937457215674)
