+++ tests/e2e/nightly/single_node/ops/singlecard_ops/triton/test_fused_recurrent_kda_npu.py
@@ -15,7 +15,7 @@ import torch.nn.functional as F
 import torch_npu  # noqa: F401
 
 from vllm_ascend.ops.triton.fla.kda import fused_recurrent_kda
-from vllm_ascend.ops.triton.fla.l2norm_kda import l2norm_fwd
+from vllm_ascend.ops.triton.fla.l2norm import l2norm_fwd
 
 DEVICE = "npu"
